updated payOffering

+++ test/byu/cit260/cityOfAaron/control/CropControlTest.java
@@ -164,8 +164,8 @@ public class CropControlTest {
         cropData.setCropYield(20);
         int offering = 10;
         //CropData cropData = null;
-        int expResult = 2.0;
-        int result = CropControl.payOffering(offering, cropData);
+        long expResult = 2;
+        long result = CropControl.payOffering(offering, cropData);
         assertEquals(expResult, result);
         // TODO review the generated test code and remove the default call to fail.
         //fail("The test case is a prototype.");
